requirements.txt
boto3==1.*
python-dotenv==1.*
pandas==2.*
pyarrow==17.*
